+++ src/Repositories/SqlStatementRepository.php
@@ -118,6 +118,8 @@ class SqlStatementRepository extends \TechDivision\Import\Repositories\SqlStatem
                 AND t1.product_id = t2.entity_id',
         SqlStatementKeys::STOCK_ITEM =>
             'SELECT * FROM ${table:cataloginventory_stock_item} WHERE product_id = :product_id AND website_id = :website_id AND stock_id = :stock_id',
+        SqlStatementKeys::STOCK_ITEM_STATUS =>
+            'SELECT * FROM ${table:cataloginventory_stock_status} WHERE product_id = :product_id AND website_id = :website_id AND stock_id = :stock_id',
         SqlStatementKeys::PRODUCT_RELATION =>
             'SELECT *
                FROM ${table:catalog_product_relation}
@@ -284,6 +286,23 @@ class SqlStatementRepository extends \TechDivision\Import\Repositories\SqlStatem
             'DELETE
                FROM ${table:catalog_product_entity_text}
               WHERE value_id = :value_id',
+        SqlStatementKeys::CREATE_STOCK_ITEM_STATUS =>
+            'INSERT 
+               INTO ${table:cataloginventory_stock_status}
+                    (${column-names:cataloginventory_stock_status})
+             VALUES (${column-placeholders:cataloginventory_stock_status})',
+        SqlStatementKeys::UPDATE_STOCK_ITEM_STATUS =>
+            'UPDATE ${table:cataloginventory_stock_status} 
+                SET ${column-values:cataloginventory_stock_status}
+              WHERE product_id = :product_id
+                AND website_id = :website_id
+                AND stock_id = :stock_id',
+        SqlStatementKeys::DELETE_STOCK_ITEM_STATUS_BY_SKU =>
+            'DELETE t1.*
+               FROM ${table:cataloginventory_stock_status} t1
+         INNER JOIN ${table:catalog_product_entity} t2
+              WHERE t2.sku = :sku
+                AND t1.product_id = t2.entity_id',
         SqlStatementKeys::CREATE_STOCK_ITEM =>
             'INSERT INTO ${table:cataloginventory_stock_item} (%s) VALUES (:%s)',
         SqlStatementKeys::UPDATE_STOCK_ITEM =>