+++ src/Repositories/StockItemRepository.php
@@ -38,6 +38,13 @@ class StockItemRepository extends AbstractRepository implements StockItemReposit
     protected $stockItemStmt;
 
     /**
+     * The prepared statement to load the existing stock items.
+     *
+     * @var \PDOStatement
+     */
+    protected $stockItemStatusStmt;
+
+    /**
      * Initializes the repository's prepared statements.
      *
      * @return void
@@ -48,6 +55,8 @@ class StockItemRepository extends AbstractRepository implements StockItemReposit
         // initialize the prepared statements
         $this->stockItemStmt =
             $this->getConnection()->prepare($this->loadStatement(SqlStatementKeys::STOCK_ITEM));
+        $this->stockItemStatusStmt =
+            $this->getConnection()->prepare($this->loadStatement(SqlStatementKeys::STOCK_ITEM_STATUS));
     }
 
     /**
@@ -72,5 +81,29 @@ class StockItemRepository extends AbstractRepository implements StockItemReposit
         // load and return the stock status with the passed product/website/stock ID
         $this->stockItemStmt->execute($params);
         return $this->stockItemStmt->fetch(\PDO::FETCH_ASSOC);
+    }
+
+    /**
+     * Load's and return's the stock item status with the passed product/website/stock ID.
+     *
+     * @param integer $productId The product ID of the stock item to load
+     * @param integer $websiteId The website ID of the stock item to load
+     * @param integer $stockId   The stock ID of the stock item to load
+     *
+     * @return array The stock item status
+     */
+    public function findOneStockStatusByProductIdAndWebsiteIdAndStockId($productId, $websiteId, $stockId)
+    {
+
+        // prepare the params
+        $params = array(
+            MemberNames::PRODUCT_ID => $productId,
+            MemberNames::WEBSITE_ID => $websiteId,
+            MemberNames::STOCK_ID   => $stockId
+        );
+
+        // load and return the stock status with the passed product/website/stock ID
+        $this->stockItemStatusStmt->execute($params);
+        return $this->stockItemStatusStmt->fetch(\PDO::FETCH_ASSOC);
     }
 }